fix: unblock clang ASAN/UBSan build with MSVC UCRT headers
Add _CRT_SECURE_NO_WARNINGS, _CRT_NONSTDC_NO_WARNINGS and
_CRT_NONSTDC_NO_DEPRECATE to cmake/VF2Warnings.cmake for non-MinGW
Windows builds (cl.exe and clang.exe against the MSVC UCRT headers).
MinGW GCC (CMAKE_C_COMPILER_ID == GNU) is unchanged and continues to
use its own non-deprecating msvcrt.

Fix the pre-existing -Wconversion warning at
src/analysis/orchestrator_scan.c:25 by wrapping the existing
(uint16_t)bytes[0] | ((uint16_t)bytes[1] << 8u) expression in an
outer (uint16_t)(...) cast, matching the canonical read_u16 pattern
in src/recovered/texture_bridge_texture.c:30.

This unblocks clang 22.1.1 with -fsanitize=address,undefined
-fno-omit-frame-pointer -Werror against the MSVC SDK: all 73 targets
compile and link cleanly and all 29 CTest tests pass with no
sanitizer violations under the dynamic
clang_rt.asan_dynamic-x86_64.dll runtime.

+++ src/analysis/orchestrator_scan.c
@@ -22,7 +22,7 @@ static vf2_status read_u16(
     if (status != VF2_OK) {
         return status;
     }
-    *value = (uint16_t)bytes[0] | ((uint16_t)bytes[1] << 8u);
+    *value = (uint16_t)((uint16_t)bytes[0] | ((uint16_t)bytes[1] << 8u));
     return VF2_OK;
 }
